feat(ui): add Run and Stop commands to the native menu and implement shortcuts feat(node_graph): enhance node dragging interaction logic feat(parallel_decoder): improve comments for clarity and remove unused variables fix(formatter): simplify filter_map usage for better readability fix(dsl_file): remove unnecessary mapping in read_packed_block method fix(sample): update sample offset check for multiple of 8 fix(waveform_index): simplify parameter passing in function call

+++ crates/dsl/src/runtime/waveform_index/reader.rs
@@ -433,7 +433,7 @@ where
                 channel,
                 start_sample,
                 end_sample,
-                target_points as usize,
+                target_points,
                 group_samples,
             )?);
         }
